Implement @transactional annotation to services.

+++ src/main/java/com/obed/retoCP2024/services/impl/OrderServiceImpl.java
@@ -2,6 +2,7 @@ package com.obed.retoCP2024.services.impl;
 
 import com.obed.retoCP2024.entities.Order;
 import com.obed.retoCP2024.entities.Product;
+import com.obed.retoCP2024.exceptions.ResourceNotFoundException;
 import com.obed.retoCP2024.repository.OrderRepository;
 import com.obed.retoCP2024.repository.ProductRepository;
 import com.obed.retoCP2024.services.OrderService;
@@ -10,11 +11,14 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional(propagation = Propagation.REQUIRED)
 public class OrderServiceImpl implements OrderService {
 
     @Autowired
@@ -23,16 +27,18 @@ public class OrderServiceImpl implements OrderService {
     private OrderRepository orderRepository;
 
     @Override
+    @Transactional
     public Order createOrder(Order order) {
 
         Product product = productRepository.findById(order.getProduct().getId())
-                .orElseThrow(() -> new IllegalArgumentException("Product not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Product not found."));
 
         order.setProduct(product);
         return orderRepository.save(order);
     }
 
     @Override
+    @Transactional
     public Order updateOrder(Long id, Order purchaseOrder) {
         Optional<Order> order = orderRepository.findById(id);
         if(!order.isPresent()) return orderRepository.save(purchaseOrder);
@@ -44,21 +50,26 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
+    @Transactional
     public void deleteOrder(Long id) {
+        if(!orderRepository.existsById(id)) throw new ResourceNotFoundException("Order not found.");
         orderRepository.deleteById(id);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Order getOrderById(Long id) {
-        return orderRepository.findById(id).orElse(null);
+        return orderRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Order not found."));
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Order> getAllOrders() {
         return orderRepository.findAll();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Page<Order> getAllOrders(Pageable pageable) {
         if(pageable == null)
             pageable = PageRequest.of(0,20);
